Move cached options into a separate .h/.cpp file
