+++ src/cached_options.cpp
@@ -1,0 +1,13 @@
+#include "cached_options.h"
+
+bool test_mode = false;
+bool debug_mode = false;
+bool use_tiles;
+bool log_from_top;
+int message_ttl;
+int message_cooldown;
+bool trigdist;
+bool fov_3d;
+int fov_3d_z_range;
+bool tile_iso;
+bool pixel_minimap_option = false;
