+++ src/game_ui.cpp
@@ -1,14 +1,5 @@
 #include "game_ui.h"
 
-#include <memory>
-
-#include "game.h"
-
-void game_ui::init_ui()
-{
-    g->init_ui( true );
-}
-
 #if !defined(TILES)
 
 void reinitialize_framebuffer()
