+++ src/popup.cpp
@@ -4,6 +4,7 @@
 #include <array>
 #include <memory>
 
+#include "cached_options.h"
 #include "catacharset.h"
 #include "ime.h"
 #include "input.h"
@@ -17,8 +18,6 @@
 #       include <SDL.h>
 #   endif
 #endif // TILES
-
-extern bool test_mode;
 
 query_popup::query_popup()
     : cur( 0 ), default_text_color( c_white ), anykey( false ), cancel( false ), ontop( false ),
